Update paper.md
I've gone through the full paper.  A lot of changes; took a lot of time.  The figure numbers are wrong.  The terms used for "chemical list" need to be consistent and there are a number of minor improvements still to make.  So, getting closer.  Please discuss with me any significant changes.  It will need at least one more iteration.