Open Source, Conclusion,...the rest
general comment: I know the importance of Global-Chem to be linked to CGenFF, but a lot of functionalities can be used beyond of that. You also wrote that the main point is to create the chemical space in chemical lists like a "free record collection". Also the title tells that.
In the Intro and Conclusion it is more focused on the force field. 
Idk if (i) the FF parametrization shouldn't rather be part of the applications,...(but I know that list selection depends on parametrization) or (ii) the title should include something FF realted, or (iii) just keep as it is ;)