review education
corrected some typos
found "inconsistencies" with spaces in chemical names: like diethyl malonate or ethyl acetate are written with a space but dimethylsulfone and others are not
some names are common names or preferred IUPAC names, some are systematic IUPAC names, should we add the others as well? I thought because this list is  also for educational purposes, that students know both or more names (or aren't confused if they are right, but it's not written there).