Introduction: few changes
Sentences could still be shortened down and made more readable. Paragraphs could be transitioned better.

In the Intro a big part is about FFs (CGenFF) and the penalty scores, but in the title it is not mentioned once. Maybe the title could be extended or reformulated (or just left as it is). I know the manuscript is about global-chem and the FF parametrization is one important application.